Refactor: extract geometry base class, unify connection control points
1. Extract DefaultNodeGeometryBase from duplicated geometry classes

   DefaultHorizontalNodeGeometry and DefaultVerticalNodeGeometry shared
   ~120 lines of identical code (constructor, boundingRect, size,
   captionRect, portTextRect, maxPortsTextAdvance, maxPortsExtent, and
   all 4 member variables). Extract these into a new intermediate base
   class DefaultNodeGeometryBase, leaving only orientation-specific
   methods in the subclasses.

2. Unify pointsC1C2Horizontal/pointsC1C2Vertical

   These two methods had identical algorithmic structure with X/Y axes
   swapped. Replace with a single computeControlPoints(Qt::Orientation)
   method that parameterizes by axis.

3. Extract selectedItemsOfType<T> template helper

   selectedNodes() and selectedGroups() shared the same iteration
   pattern over QGraphicsScene::selectedItems() with qgraphicsitem_cast.
   Extract into a local template function.
